Ignoring JVM warning in pdfparser
This app is deployed to Heroku. Heroku sets the JAVA_TOOL_OPTIONS env
variable. When we run Java on the command line, it issues a warning to
stderr, which was leading to 500 errors when attempting to download the
SF PDF. Since this isn't a real error, we ignored it by filtering.

stackoverflow:
https://stackoverflow.com/questions/11683715/suppressing-the-picked-up-java-options-message

heroku setting JAVA_TOOL_OPTIONS variable
https://devcenter.heroku.com/articles/java-support#adjusting-environment-for-a-dyno-size

Signed-off-by: Christa Hartsock <christa@codeforamerica.org>
Co-authored-by: Christa Hartsock <christa@codeforamerica.org>