Prevent early exit for failed notification send
Previously if an organization didn't have an email address for a profile
for some reason (was happening on staging for some counties we hadnt' yet onboarded),
the job to send unread application notifications for the day would bail
out early, preventing some orgs from receiving notifications.

This commit attempts to fix that by catching the raised exception,
collecting it in case there are multiple, and re-raising it at the end
of the task.