Allow user to soft delete clips
So that we can clean up the clips page.

This should do a soft delete. In practice it looks like we've deleted,
but individual clips can still be accessed in the database if we need
them. FOr now, we haven't put an easy GUI around this, though, so would
require directly modifying the database.

[Finishes #173345079]