Unset the max number of retries for Celery
Currently our site is down because AMPQ, the service that stores our
Celery queueus, is down. When our app tries to save a task to it, it
blows up. This attempts to remove the cap on the number or retries so
that it no longer blows up, and instead just waits until the service is
available.

Potentially dangerous? Maybe!

[#172236152]