Increase session cookie expiration
We're seeing user sessions get dropped on mobile, and we're not totally
clear on what's happening but suspect that clients might be taking
longer than the session expiration to complete an application, which
results in them being redirected to the home page and losing their app.

This is us trying to see if extending the session cookie expiration
helps—intend to check in later and see.

[#171547782]